requirements.txt
Django==1.10.6
django-debug-toolbar
Pillow
django-ckeditor==5.2.1
django-content-editor==1.0.0
django-versatileimagefield
--upgrade pip
django-simple-captcha
django-mptt
